+++ backend/app/models/inventory.py
@@ -1,12 +1,12 @@
-from sqlalchemy import Column, BigInteger, Integer, String, Text, SmallInteger, Date, DateTime, ForeignKey, func
+from sqlalchemy import Column, Integer, String, Text, SmallInteger, Date, DateTime, ForeignKey, func
 from app.database import Base
 
 
 class InventoryCheck(Base):
     __tablename__ = "museum_inventory_check"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    room_id = Column(BigInteger, ForeignKey("museum_storage_room.id"), nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    room_id = Column(Integer, ForeignKey("museum_storage_room.id"), nullable=False)
     check_date = Column(Date, nullable=False)
     total_count = Column(Integer, default=0)
     checked_count = Column(Integer, default=0)
@@ -21,9 +21,9 @@ class InventoryCheck(Base):
 class CollectionMovement(Base):
     __tablename__ = "museum_collection_movement"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    collection_id = Column(BigInteger, ForeignKey("museum_collection.id"), nullable=False)
-    room_id = Column(BigInteger, ForeignKey("museum_storage_room.id"))
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    collection_id = Column(Integer, ForeignKey("museum_collection.id"), nullable=False)
+    room_id = Column(Integer, ForeignKey("museum_storage_room.id"))
     movement_type = Column(SmallInteger, nullable=False, comment="1入库 2出库 3移库")
     reason = Column(String(200))
     operator = Column(String(50))
